+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project are documented in this file.
 
-## [0.8.12] - Unreleased
+## [0.8.12] - 2026-09-24
 
 ### Changed